% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectNA.R
\name{detectNA}
\alias{detectNA}
\title{Detect Missing Values}
\usage{
detectNA(data)
}
\arguments{
\item{data}{A data frame or tibble}
}
\value{
A named vector with counts of missing values for each column.
}
\description{
This function detects missing values in each column of a dataset.
It returns a named vector with counts of missing values for each column that has missing data.
}
\examples{
detectNA(airquality)

}
