% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_columns.R
\name{drop_na_columns}
\alias{drop_na_columns}
\title{Drop columns with too many missing values}
\usage{
drop_na_columns(df, threshold = 50)
}
\arguments{
\item{df}{A data frame}

\item{threshold}{descriptionA numeric value between 0 and 100 that sets the
maximum allowed percentage of missing values. Columns that have a higher
percentage of missing values will be dropped.}
}
\value{
A data frame with columns dropped if they exceed the threshold
}
\description{
This function removes columns from the input data frame if the percentage of
missing values in that column exceeds the input threshold.
}
\examples{
df <- data.frame(a = c(1, NA, 3),
                 b = c(NA, NA, NA),
                 c = c(2, 4, NA))
drop_na_columns(df, threshold = 50)

}
