% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_na.R
\name{extract_na}
\alias{extract_na}
\title{Extract rows from a data frame that contain NA values}
\usage{
extract_na(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A data frame that contains only the rows from the input data frame that
contained NA values.
}
\description{
This function takes a data frame as input and subsets it so only rows that have
NA values are returned.
}
\examples{
df <- data.frame(a = c(1, NA, 3),
                 b = c("x", NA, "z"),
                 c = c(NA, NA, 5))
extract_na(df)

}
