% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_check.R
\name{missing_check}
\alias{missing_check}
\title{Check missing values in a data frame}
\usage{
missing_check(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
A one-row data frame with the count of missing values in each column
}
\description{
This function takes in a data frame and returns a one-row data frame with each
column corresponding to the same column in the input data frame, and the values
are the number of NA values.
}
\examples{
df <- data.frame(a = c(1, NA, 3),
                 b = c("x", NA, "z"),
                 c = c(NA, NA, 5))
missing_check(df)

}
