% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_tilemap.R
\name{missing_tilemap}
\alias{missing_tilemap}
\title{Create a tilemap highlighting missing values}
\usage{
missing_tilemap(df, main_title = "Missingness Tile Map")
}
\arguments{
\item{df}{A dataframe}

\item{main_title}{A character string giving the plot a title}
}
\value{
A ggplot object representing the missingness tile map.
}
\description{
This function creates a tilemap that shows missing values of a dataframe by
column and with the location of the row.
}
\examples{
#Using built-in airquality dataset
data("airquality")
missing_tilemap(airquality)

}
