% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_column_replace.R
\name{na_column_replace}
\alias{na_column_replace}
\title{Replace NAs in Column Function}
\usage{
na_column_replace(df, replacements)
}
\arguments{
\item{df}{The dataframe for which NA values will be replaced.}

\item{replacements}{The column name and value you wish to replace NAs with i.e. col1 = 0.}
}
\value{
A data frame with NA values replaced in specified columns
}
\description{
This function allows you to replace NA values in any columns with a specified value
}
\examples{
df <- data.frame(a = c(1, 2, NA),
                 b = c(NA, 5, 6),
                 c = c(7, NA, 9))
na_column_replace(df, c(a = 3, b = 4, c = 8))

}
