% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NATools-package.R
\docType{_PACKAGE}
\name{NATools}
\alias{NATools}
\title{NATools: A package for handling missing data}
\description{
The NATools package provide functions to detect, and handle missing values in the datasets.
}
\section{Functions}{

- `detectNA()` : Identify missing values in a dataset.
- `repNAmed()` : Replace missing values with column median.
}

