% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repNAmed.R
\name{repNAmed}
\alias{repNAmed}
\title{Replace Missing Values with Medians}
\usage{
repNAmed(dataset)
}
\arguments{
\item{dataset}{A data frame or tibble}
}
\value{
A data frame with missing values replaced by column medians.
}
\description{
This function replaces missing values in numeric columns with the median of each column.
}
\examples{
repNAmed(airquality)

}
